+++ .build/run_linux.sh
@@ -3,10 +3,13 @@ set -e
 
 echo $PWD
 
+export FIREFOX_WD_URL=http://localhost:4444/wd/hub
+export CHROME_WD_URL=http://localhost:4445/wd/hub
+
 npm ci
 npm run build
 npm run coverage:clean
-FIREFOX_WD_URL=http://localhost:4444/wd/hub CHROME_WD_URL=http://localhost:4445/wd/hub npm test -- --runInBand --ci --bail
+npm test -- --runInBand --ci --bail
 npm run test:it
 npm run coverage:merge
 npm run coverage:merge-report